refactor: issue JDBC 쓰기와 삭제 흐름 분리
